software/src/results-summary/requirements.txt
polars-lts-cpu==1.33.1
numpy==2.2.6
